+++ app/src/main/java/com/team19/animation/ui/DessertsFragment.java
@@ -1,7 +1,7 @@
-package com.team19.animation;
-
+package com.team19.animation.ui;
 
 import android.os.Bundle;
+
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.LinearLayoutManager;
@@ -10,45 +10,42 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.team19.animation.R;
 import com.team19.animation.adapters.MainScreenRecyclerAdapter;
 import com.team19.animation.models.Food;
 import com.team19.animation.utils.Foods;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 
 
-public class DrinksFragment extends Fragment {
-    private static final String TAG = "DrinksFragment";
+public class DessertsFragment extends Fragment {
+    private static final String TAG = "DessertsFragment";
     private RecyclerView recyclerView;
-    private ArrayList<Food> foodlist;
+    private ArrayList<Food> drinklist;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_drinks, container,false);
-        recyclerView = view.findViewById(R.id.drinks_recyclerview);
-        MainScreenRecyclerAdapter mainScreenRecyclerAdapter = new MainScreenRecyclerAdapter(getContext(),foodlist);
+
+        View view = inflater.inflate(R.layout.fragment_desserts, container,false);
+        recyclerView = view.findViewById(R.id.desserts_recyclerview);
+        MainScreenRecyclerAdapter mainScreenRecyclerAdapter = new MainScreenRecyclerAdapter(getContext(),drinklist);
         recyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
         recyclerView.setAdapter(mainScreenRecyclerAdapter);
         return view;
-    }
 
+    }
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        super.onCreate(savedInstanceState);
 
-        Foods food = new Foods();
-        foodlist = new ArrayList<>();
+        drinklist = new ArrayList<>();
+        Food[] desserts = new Foods().getDesserts();
 
-        for (Food drinks:food.drinks) {
-            foodlist.add(drinks);
-        }
-//        foodlist = new ArrayList<>();
-//        Food[] desserts = new Foods().getMainDishes();
-//
-//        foodlist.addAll(Arrays.asList(desserts));
-//    }
+        drinklist.addAll(Arrays.asList(desserts));
     }
 }
+
+
