+++ app/src/main/java/com/team19/animation/ui/MainDishesFragment.java
@@ -1,7 +1,7 @@
-package com.team19.animation;
+package com.team19.animation.ui;
+
 
 import android.os.Bundle;
-
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.LinearLayoutManager;
@@ -10,6 +10,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.team19.animation.R;
 import com.team19.animation.adapters.MainScreenRecyclerAdapter;
 import com.team19.animation.models.Food;
 import com.team19.animation.utils.Foods;
@@ -18,33 +19,34 @@ import java.util.ArrayList;
 import java.util.Arrays;
 
 
-public class DessertsFragment extends Fragment {
-    private static final String TAG = "DessertsFragment";
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class MainDishesFragment extends Fragment {
     private RecyclerView recyclerView;
-    private ArrayList<Food> drinklist;
+    private ArrayList<Food> foodlist;
+
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-
-        View view = inflater.inflate(R.layout.fragment_desserts, container,false);
-        recyclerView = view.findViewById(R.id.desserts_recyclerview);
-        MainScreenRecyclerAdapter mainScreenRecyclerAdapter = new MainScreenRecyclerAdapter(getContext(),drinklist);
+        // Inflate the layout for this fragment
+        View view = inflater.inflate(R.layout.fragment_main_dishes, container,false);
+        recyclerView = view.findViewById(R.id.maindishes_recyclerview);
+        MainScreenRecyclerAdapter mainScreenRecyclerAdapter = new MainScreenRecyclerAdapter(getContext(),foodlist);
         recyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
         recyclerView.setAdapter(mainScreenRecyclerAdapter);
         return view;
-
     }
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        drinklist = new ArrayList<>();
-        Food[] desserts = new Foods().getDesserts();
+        foodlist = new ArrayList<>();
+        Food[] mainDishes = new Foods().getMainDishes();
 
-        drinklist.addAll(Arrays.asList(desserts));
+        foodlist.addAll(Arrays.asList(mainDishes));
     }
+
 }
-
-
