+++ app/src/main/res/layout/fragment_desserts.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DessertsFragment">
+    tools:context=".ui.DessertsFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/desserts_recyclerview"
