+++ app/src/main/res/layout/fragment_drinks.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DrinksFragment">
+    tools:context=".ui.DrinksFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/drinks_recyclerview"
